refactor: 모델 패키지 분리
모델에 속한 클래스의 개수가 많아서
purchase, draw, result 3개의 하위 패키지로 분리함.
