+++ src/main/java/lotto/model/purchase/Purchase.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.purchase;
 
 import static lotto.enums.LottoConstant.PRICE;
