+++ src/test/java/lotto/model/LottoNumbersTest.java
@@ -3,6 +3,7 @@ package lotto.model;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import lotto.model.draw.LottoNumbers;
 import org.junit.jupiter.api.Test;
 
 class LottoNumbersTest {
