+++ src/test/java/lotto/model/LottoResultTest.java
@@ -5,6 +5,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.EnumMap;
 import java.util.List;
 import lotto.enums.LottoWinInfo;
+import lotto.model.draw.DrawNumber;
+import lotto.model.draw.LottoNumbers;
+import lotto.model.purchase.Lotto;
+import lotto.model.result.LottoResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
