+++ src/test/java/lotto/model/ProfitTest.java
@@ -9,6 +9,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.EnumMap;
 import lotto.enums.LottoWinInfo;
+import lotto.model.result.Profit;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
